requirements.txt
tensorflow==2.1.0
Django==3.1.5
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
numpy==1.19.5
opencv-python==4.5.1.48
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.4.1
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wrapt==1.12.1
zipp==3.4.0
cassandra-driver==3.24.0
